requirements.txt
pandas
yfinance
sqlalchemy
google-genai
pyodbc
python-dotenv